转移等级问题、幸运值问题、护甲恢复

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/buffs/机制/连杀状态.java
@@ -23,7 +23,7 @@ public class 连杀状态 extends FlavourBuff{
         return count+super.name();
     }@Override
     	public String iconTextDisplay() {
-    		return count+"";
+    		return (int)(count)+"";
     	}
 
     public float count = 0;
